Post y Comment

+++ proyectos/large/views/post/see_post.php
@@ -1,25 +1,31 @@
-<div role="group">
-    <button><a class="button-action" href="<?= base_url ?>post/edit&id=<?= $post->id ?>">Editar</a></button>
-    <button><a class="button-action" href="<?= base_url ?>post/delete&id=<?= $post->id ?>">Eliminar</a></button>
-    <button
-        <?php if ($post->status == 'Visible') : ?>
-        disabled
-        <?php endif; ?>>
-        <a class="button-action" href="<?= base_url ?>post/status&id=<?= $post->id ?>&status=Visible">
-            Visible
-        </a>
-    </button>
+<!-- __<DIV> DE BOTONES DE EDITAR, ELIMINAR, VISIBLE Y OCULTAR__ -->
+<!-- _ESTE <DIV> INTERACTIVO SOLO SE MOSTRARÁ SI SE ESTÁ LOGEADO Y EL POST PERTENECE AL USUARIO REGISTRADO_ -->
+<?php if (isset($_SESSION['identity']) && $_SESSION['identity']->id == $post->user_id): ?>
+    <div role="group">
 
-    <button
-        <?php if ($post->status == 'Oculto') : ?>
-        disabled
-        <?php endif; ?>>
-        <a class="button-action" href="<?= base_url ?>post/status&id=<?= $post->id ?>&status=Oculto">
-            Ocultar
-        </a>
-    </button>
+        <button><a class="button-action" href="<?= base_url ?>post/edit&id=<?= $post->id ?>">Editar</a></button>
+        <button><a class="button-action" href="<?= base_url ?>post/delete&id=<?= $post->id ?>">Eliminar</a></button>
+        <button
+            <?php if ($post->status == 'Visible') : ?>
+            disabled
+            <?php endif; ?>>
+            <a class="button-action" href="<?= base_url ?>post/status&id=<?= $post->id ?>&status=Visible">
+                Visible
+            </a>
+        </button>
 
-</div>
+        <button
+            <?php if ($post->status == 'Oculto') : ?>
+            disabled
+            <?php endif; ?>>
+            <a class="button-action" href="<?= base_url ?>post/status&id=<?= $post->id ?>&status=Oculto">
+                Ocultar
+            </a>
+        </button>
+
+    </div>
+<?php endif; ?>
+
 
 <!-- __ALERTA DE CAMBIO DE STATUS__ -->
 <?php if (isset($_SESSION['post-status'])) : ?>
@@ -50,63 +56,139 @@
     <?php unset($_SESSION['post-update']); ?>
 <?php endif; ?>
 
+
+<!-- __VISUALIZACIÓN DEL POST__ -->
 <h2 class="title-category"><?= $post->title ?></h2>
 
-<?php $categoryClass = strtolower(str_replace(' ', '', $post->category_name)); ?><!--  utilizar str_replace -->
+<?php $categoryClass = strtolower(str_replace(' ', '', $post->category_name)); ?>
 <?php $statusClass = strtolower($post->status); ?>
-<a href="<?= base_url ?>post/see&id=<?= $post->id ?>" class="post_link <?= $categoryClass; ?>">
-    <article class="post <?= $post->status == 'Oculto' ? $statusClass : $categoryClass ?>">
+
+<article class="post <?= $post->status == 'Oculto' ? $statusClass : $categoryClass ?>">
 
 
-        <section class="post_head">
-            <?php if ($post->status == 'Ocultar') : ?>
+    <section class="post_head">
 
-                <div>
-                    <i class="fa-solid fa-eye-slash"></i> Post Oculto
-                </div>
+        <?php if ($post->status == 'Ocultar') : ?>
 
-            <?php endif; ?>
-            <p class="category_post"><?= $post->category_name ?></p>
-            <p class="post_sub_title"><?= $post->sub_title ?></p>
-            <p class="post_detail">Publicado el: <?= date('Y-m-d H:i', strtotime($post->created_at)) ?> by <?= $post->creator ?></p>
-        </section>
-        <p><?= nl2br(htmlspecialchars($post->content)) ?></p>
-        <?php if ($post->image != null) : ?>
-            <img class="post_image" src="<?= base_url ?>uploads/images/<?= $post->image ?>" alt="">
+            <div>
+                <i class="fa-solid fa-eye-slash"></i> Post Oculto
+            </div>
+
         <?php endif; ?>
+        <p class="category_post"><?= $post->category_name ?></p>
+        <p class="post_sub_title"><?= $post->sub_title ?></p>
+        <p class="post_detail">Publicado el: <?= date('Y-m-d H:i', strtotime($post->created_at)) ?> by <?= $post->creator ?></p>
 
-        <!-- <div class="likes_coments">
-            <div>
-                <i class="fa-solid fa-thumbs-up"></i>
-            </div>
-            <div>
-                <i class="fa-solid fa-thumbs-down"></i>
-            </div>
-            <div>
-                <i class="fa-solid fa-comments"></i>
-            </div>
-        </div> -->
+    </section>
 
-        <div role="group" class="post_interaction">
-            <button data-tooltip="Like"><i class="fa-regular fa-thumbs-up"></i> 52</button>
-            <button data-tooltip="Dislike"><i class="fa-regular fa-thumbs-down"></i> 4</button>
-            <button data-tooltip="Comentar"><i class="fa-regular fa-comments"></i> 7</button>
-            <button data-tooltip="Compartir"> <i class="fa-regular fa-share-from-square"></i></button>
+    <p><?= nl2br(htmlspecialchars($post->content)) ?></p>
 
+    <?php if ($post->image != null) : ?>
+        <img class="post_image" src="<?= base_url ?>uploads/images/<?= $post->image ?>" alt="">
+    <?php endif; ?>
+
+
+    <!-- Botón de Comentar en views/post/see_post.php -->
+    <div role="group" class="post_interaction">
+
+        <button data-tooltip="Like"><i class="fa-regular fa-thumbs-up"></i> 52</button>
+
+        <button data-tooltip="Dislike"><i class="fa-regular fa-thumbs-down"></i> 4</button>
+
+        <button data-target="comment-modal" onclick="toggleModal(event)" data-tooltip="Comentar">
+            <i class="fa-regular fa-comments"></i>
+            <?= $commentCount ?>
+        </button>
+
+        <button data-tooltip="Compartir"> <i class="fa-regular fa-share-from-square"></i></button>
+
+    </div>
+
+    <?php if (isset($_SESSION['identity'])) : ?>
+
+        <form action="<?= base_url ?>comment/save" method="POST">
+
+            <input type="hidden" name="post_id" value="<?= $post->id; ?>">
+            <textarea placeholder="Escribe tu comentario aquí..." name="content"></textarea>
+            <input type="submit" value="Comentar">
+
+            <!-- <button autofocus data-target="comment-modal" onclick="toggleModal(event)">Comentar</button> -->
+            </footer>
+        </form>
+
+    <?php else : ?>
+
+        <div>
+            Si deseas comentar este post, es necesario iniciar sesión.
         </div>
 
-        <section class="comentarios">
-       
+    <?php endif; ?>
 
-            <?php if (isset($allComments) && $allComments->num_rows > 0): ?>
-                <?php while ($comment = $allComments->fetch_object()): ?>
-                    <div class="comentario">
-                        <p><?= $comment->content ?></p>
-                    </div>
-                <?php endwhile; ?>
-            <?php else: ?>
-                <p>No hay comentarios disponibles.</p>
-            <?php endif; ?>
-        </section>
-    </article>
-</a>
+    <!-- Modal para Comentar -->
+    <dialog id="comment-modal">
+
+        <article>
+
+            <header>
+                <button aria-label="Close" rel="prev" data-target="comment-modal" onclick="toggleModal(event)"></button>
+                <h3>Deja tu comentario</h3>
+            </header>
+
+            <form action="<?= base_url ?>comment/save" method="POST">
+
+                <p>
+                    <input type="hidden" name="post_id" value="<?= $post->id; ?>">
+                    <textarea placeholder="Escribe tu comentario aquí..." name="content"></textarea>
+                </p>
+
+                <footer class="modal_footer">
+                    <input type="submit" value="Comentar" data-target="comment-modal" onclick="toggleModal(event)">
+                    <button role="button" class="secondary" data-target="comment-modal" onclick="toggleModal(event)">Cancelar</button>
+                    <!-- <button autofocus data-target="comment-modal" onclick="toggleModal(event)">Comentar</button> -->
+                </footer>
+
+            </form>
+
+        </article>
+
+    </dialog>
+
+
+    <!-- __ALERTA DE CRECIÓN DE COMENTARIOS__ -->
+    <?php if (isset($_SESSION['register-comment'])) : ?>
+        <?php if ($_SESSION['register-comment'] == 'complete') : ?>
+            <div class="alert alert-success">
+                <i class="fa-regular fa-square-check"></i> Comentario creado con extio!
+            </div>
+        <?php elseif ($_SESSION['register-comment'] == 'falied') : ?>
+            <div class="alert alert-danger">
+                <i class="fa-solid fa-xmark"></i> Error hacer el comentario, por favor intentalo nuevamente.
+            </div>
+        <?php endif; ?>
+        <?php unset($_SESSION['register-comment']); ?>
+    <?php endif; ?>
+
+
+
+    <!-- __SECCIÓN DE COMENTARIOS__ -->
+    <section class="container-comments">
+
+        <?php if (isset($allComments) && $allComments->num_rows > 0): ?>
+
+            <?php while ($comment = $allComments->fetch_object()): ?>
+
+                <div class="comments">
+
+                    <p class="creator_date"><?= $comment->creator ?> • <?= Utils::timeAgo($comment->created_at) ?></p>
+                    <p class="comment"><?= $comment->comment ?></p>
+
+                </div>
+
+            <?php endwhile; ?>
+
+        <?php else: ?>
+            <p>No hay comentarios disponibles.</p>
+        <?php endif; ?>
+
+    </section>
+</article>
